Add files via upload
This one was a little bit frustrating. I didn't think about combining the two names in the beginning and just created two for loops that run each name. For some reason, the result was incorrect. Combining strings was key for this problem since we only need to count the letters. I need to pay more attention to what I can and cannot do.