*nested-list
big_list[nested_list][nested_list_index_value] - this is the order to call a value from a specific list.
This data structure took me A LOT of thinking to fully understand it. The concept of nested-list becomes quite confusing when we have to change values so that normal people can interact with our programs. I had to constantly remind myself that it is the function that we're working with; naming a variable is to assist our understanding, it does not change how a function operates. This one was a difficult one. I need to review it from time to time.