*
This was my first encounter with a high-difficulty project. This project required so much thinking, and trial and error. I learned important lessons from this project. I should focus on creating a working code before adding more features to it. In the beginning, I tried to satisfy all the required conditions such as having 6 lives or warning the user this and that; eventually, it makes my code become messy and I hit a dead-end with no way to progress. I need to break things down into small working blocks of code, trying to do multiple things at once got me nowhere.