*revisit from time to time
This was my first encounter with a high-difficulty project. This project required so much thinking, and trial and error. I learned a few important lessons from this project.
In the beginning, I tried to satisfy all the required conditions such as having 6 lives or warning the user this and that; eventually, it makes my code become messy and I hit a dead-end with no way to progress. I should focus on creating a working code before adding more features to it. Don't worry about  UX, it can wait.
I need to focus on breaking things down into small, working blocks of code, doing multiple things at once got me nowhere.