Add files via upload
This program was difficult. It requires a change from the traditional way of thinking. My brain has been wired to always look for 'TRUE', so I spent too much time making my 'TRUE' arguments become the primary 'if' statement. Satisfying the conditions that make a prime number 'FALSE' is a lot easier. I need to think about this from now on.
If making something become 'TRUE' is too difficult, maybe the 'FALSE' approach will be easier.