*fun, hard, learned a lot
This program was really hard. Everything was ok until the 'shift_number' is out-of-bound. Originally, all I could think of was to increase the size of the alphabet, but that solution will restrict the 'shift_number' which I think is an unsatisfying solution; plus it will greatly increase the size of the list just by repeating its value.

I had a talk with my sister and she suggested that I could use modulus if I want the 'shift_number' to stay relative to the original list no matter how great the 'shift_number' is. This was such a great idea. I struggled a lot to find the correct way to implement the idea but eventually got it.

However, now that I understand a little bit more about programming, I'm sure a cipher of this caliber will be cracked within seconds by a loop that checks for all possible combination from 0 to infinity. Still, this was such a fun program to make. Really enjoy making this program.