*nested dictionary
This program was relatively easy to do. While complicated at first, the concept of dictionary and nested dictionary is actually quite intuitive. It's like a building, and each key has access to a specific room within a building. This data structure is advantageous to store data that comes in pairs and needs to be under a specific label.

To access a dictionary, always specify the key, not the value. If there's a list nested within, specify the index after the key.
order = {
    "starter": {1: "Salad", 2: "Soup"},
    "main": {1: ["Burger", "Fries"], 2: ["Steak"]},
    "dessert": {1: ["Ice Cream"], 2: []},
}
To print "Steak" we call, order["main"][2][0] (order[main_key][sub_key][index_0])