*functions, outputs, recursion
I learn how to create my own functions for a program. I also learn the difference between 'print' and 'return'. 'print' is simply to display whereas 'return' returns a value in order to exit a function. Additionally, 'return' allows the value to be used in other functions.
'Recursion' is a new concept. It allows a function to call itself. This is useful when we need a function to run continuously while predefined conditions are met. Without a terminating condition, 'recursion' will cause an infinite loop.